% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{calc_cc_allcases_by_age_acq}
\alias{calc_cc_allcases_by_age_acq}
\title{Tabulate cases with condition in causal chain by age and site of acquisition}
\usage{
calc_cc_allcases_by_age_acq(d, condition, age_groups = NULL, sites = NULL)
}
\arguments{
\item{d}{A data object obtained from \code{\link{load_data}}.}

\item{condition}{A string specifying the condition to tabulate causal chain
presence for. To list all possibilities, see
\code{\link{valid_conditions}}.}

\item{age_groups}{Optional vector of age groups.
See \code{\link{valid_age_groups}}.}

\item{sites}{An optional vector of site names to include in the tabulation.
If not provided, all sites will be used. See \code{\link{valid_sites}}.}
}
\description{
Tabulate cases with condition in causal chain by age and site of acquisition
}
\examples{
calc_cc_allcases_by_age_acq(mock,
  condition = "Klebsiella pneumoniae",
  age_groups = valid_age_subcats(mock)
)
}
