% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{calc_cc_allcases_by_site_age}
\alias{calc_cc_allcases_by_site_age}
\title{Tabulate the number of cases of a given condition found in the causal chain
  by site and by age, within the context of all cases}
\usage{
calc_cc_allcases_by_site_age(d, condition, sites = NULL)
}
\arguments{
\item{d}{A data object obtained from \code{\link{load_data}}.}

\item{condition}{A string specifying the condition to tabulate causal chain
presence for. To list all possibilities, see
\code{\link{valid_conditions}}.}

\item{sites}{An optional vector of site names to include in the tabulation.
If not provided, all sites will be used. See \code{\link{valid_sites}}.}
}
\description{
Tabulate the number of cases of a given condition found in the causal chain
  by site and by age, within the context of all cases
}
\examples{
calc_cc_allcases_by_site_age(mock, condition = "Streptococcus agalactiae")
}
