% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{calc_cc_detected_by_site_age}
\alias{calc_cc_detected_by_site_age}
\title{Tabulate the number of cases of a given condition found in the causal chain
  by site and by age, within the context of cases where the pathogen(s)
  are detected in TAC results (dying from vs. dying with)}
\usage{
calc_cc_detected_by_site_age(
  d,
  condition,
  pathogen,
  sites = NULL,
  specimen_types = NULL
)
}
\arguments{
\item{d}{A data object obtained from \code{\link{load_data}}.}

\item{condition}{A string specifying the condition to tabulate causal chain
presence for. To list all possibilities, see
\code{\link{valid_conditions}}.}

\item{pathogen}{A string specifying the pathogen to count positive cases of
in the TAC results.}

\item{sites}{An optional vector of site names to include in the tabulation.
If not provided, all sites will be used. See \code{\link{valid_sites}}.}

\item{specimen_types}{An optional vector of specimen types to include in the
calculation. If not provided, all specimen types will be used. See
\code{\link{valid_specimen_types}}.}
}
\description{
Tabulate the number of cases of a given condition found in the causal chain
  by site and by age, within the context of cases where the pathogen(s)
  are detected in TAC results (dying from vs. dying with)
}
\examples{
specimen_types <- c(
  "Cerebrospinal fluid sample",
  "Tissue specimen from lung",
  "Whole blood",
  "Rectal swab",
  "Plasma or spun blood specimen"
)

calc_cc_detected_by_site_age(mock,
  condition = "Streptococcus agalactiae",
  pathogen = "Group B Streptococcus",
  specimen_types = specimen_types)
}
