% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{calc_nspecimen_by_pmi}
\alias{calc_nspecimen_by_pmi}
\title{Tabulate the number of positive specimens for each case where the pathogen(s)
  are detected in TAC results by postmortem interval range (PMI -
  time from death to MITS)}
\usage{
calc_nspecimen_by_pmi(d, pathogen, sites = NULL)
}
\arguments{
\item{d}{A data object obtained from \code{\link{load_data}}.}

\item{pathogen}{A string specifying the pathogen to count positive cases of
in the TAC results.}

\item{sites}{An optional vector of site names to include in the tabulation.
If not provided, all sites will be used. See \code{\link{valid_sites}}.}
}
\description{
Tabulate the number of positive specimens for each case where the pathogen(s)
  are detected in TAC results by postmortem interval range (PMI -
  time from death to MITS)
}
\examples{
calc_nspecimen_by_pmi(mock, pathogen = "Group B Streptococcus")
}
