% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{calc_pmi_by_decode_site}
\alias{calc_pmi_by_decode_site}
\title{Calculate average postmortem interval (PMI - average time from death to MITS
  in hours) by DeCoDe result and site}
\usage{
calc_pmi_by_decode_site(
  d,
  condition,
  pathogen,
  icds,
  specimen_types = NULL,
  sites = NULL
)
}
\arguments{
\item{d}{A data object obtained from \code{\link{load_data}}.}

\item{condition}{A string specifying the condition to tabulate causal chain
presence for. To list all possibilities, see
\code{\link{valid_conditions}}.}

\item{pathogen}{A string specifying the pathogen to count positive cases of
in the TAC results.}

\item{icds}{A vector of ICD10 codes to check for the DeCoDe result of
"Contributing (P2)".}

\item{specimen_types}{An optional vector of specimen types to include in the
calculation. If not provided, all specimen types will be used. See
\code{\link{valid_specimen_types}}.}

\item{sites}{An optional vector of site names to include in the tabulation.
If not provided, all sites will be used. See \code{\link{valid_sites}}.}
}
\description{
Calculate average postmortem interval (PMI - average time from death to MITS
  in hours) by DeCoDe result and site
}
\examples{
specimen_types <- c(
  "Cerebrospinal fluid sample",
  "Tissue specimen from lung",
  "Whole blood",
  "Rectal swab",
  "Plasma or spun blood specimen"
)

calc_pmi_by_decode_site(mock,
  condition = "Streptococcus agalactiae",
  pathogen = "Group B Streptococcus",
  icds = c("P36.0", "A40.1", "P23.3", "G00.2"),
  specimen_types = specimen_types)
}
