% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{calc_pmi_by_specimen_site}
\alias{calc_pmi_by_specimen_site}
\title{Calculate average postmortem interval (PMI - average time from death to MITS
  in hours) by specimen type of NP Only vs. Blood/CSF/Lung and by site}
\usage{
calc_pmi_by_specimen_site(d, pathogen, sites = NULL)
}
\arguments{
\item{d}{A data object obtained from \code{\link{load_data}}.}

\item{pathogen}{A string specifying the pathogen to count positive cases of
in the TAC results.}

\item{sites}{An optional vector of site names to include in the tabulation.
If not provided, all sites will be used. See \code{\link{valid_sites}}.}
}
\description{
Calculate average postmortem interval (PMI - average time from death to MITS
  in hours) by specimen type of NP Only vs. Blood/CSF/Lung and by site
}
\examples{
calc_pmi_by_specimen_site(mock, pathogen = "Group B Streptococcus")
}
