% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{calc_syndrome_combinations}
\alias{calc_syndrome_combinations}
\title{Tabulate syndrome combinations for a specified condition}
\usage{
calc_syndrome_combinations(
  d,
  condition,
  syndrome_names,
  syndrome_values = NULL,
  specimen_types = NULL,
  sites = NULL
)
}
\arguments{
\item{d}{A data object obtained from \code{\link{load_data}}.}

\item{condition}{A string specifying the condition to tabulate causal chain
presence for. To list all possibilities, see
\code{\link{valid_conditions}}.}

\item{syndrome_names}{A vector specifying conditions to tabulate
combinations of presence of for the given \code{condition}}

\item{syndrome_values}{An optional vector of the same length of
\code{syndrome_names} that specifies the mapping from the official
condition names to the values to use for reporting. See example.}

\item{specimen_types}{An optional vector of specimen types to include in the
calculation. If not provided, all specimen types will be used. See
\code{\link{valid_specimen_types}}.}

\item{sites}{An optional vector of site names to include in the tabulation.
If not provided, all sites will be used. See \code{\link{valid_sites}}.}
}
\description{
Tabulate syndrome combinations for a specified condition
}
\examples{
calc_syndrome_combinations(mock,
  condition = "Streptococcus pneumoniae",
  syndrome_names = c(
    "Lower respiratory infections",
    "Meningitis/Encephalitis",
    "Neonatal sepsis",
    "Congenital infection"),
  syndrome_values = c(
    "Pneumonia",
    "Meningitis",
    "Sepsis",
    "Sepsis"),
  specimen_types = c(
    "Cerebrospinal fluid sample",
    "Tissue specimen from lung",
    "Whole blood")
)
}
