% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{calc_top_dcd_pathogens_by_acq}
\alias{calc_top_dcd_pathogens_by_acq}
\title{Tabulate top pathogens associated with a condition by site of acquisition}
\usage{
calc_top_dcd_pathogens_by_acq(d, condition, age_groups = NULL, sites = NULL)
}
\arguments{
\item{d}{A data object obtained from \code{\link{load_data}}.}

\item{condition}{A string specifying the condition to tabulate causal chain
presence for. To list all possibilities, see
\code{\link{valid_conditions}}.}

\item{age_groups}{Optional vector of age groups.
See \code{\link{valid_age_groups}}.}

\item{sites}{An optional vector of site names to include in the tabulation.
If not provided, all sites will be used. See \code{\link{valid_sites}}.}
}
\description{
Tabulate top pathogens associated with a condition by site of acquisition
}
\examples{
calc_top_dcd_pathogens_by_acq(mock,
  condition = "Lower respiratory infections",
  age_groups = c(
    "Infant (28 days to less than 12 months)",
    "Child (12 months to less than 60 Months)")
)
}
