% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{calc_top_etiol_by_age}
\alias{calc_top_etiol_by_age}
\title{Tabulate top etiologies by specified age groups}
\usage{
calc_top_etiol_by_age(d, age_groups = NULL, sites = NULL)
}
\arguments{
\item{d}{A data object obtained from \code{\link{load_data}}.}

\item{age_groups}{Optional vector of age groups.
See \code{\link{valid_age_groups}}.}

\item{sites}{An optional vector of site names to include in the tabulation.
If not provided, all sites will be used. See \code{\link{valid_sites}}.}
}
\description{
Tabulate top etiologies by specified age groups
}
\examples{
calc_top_etiol_by_age(mock,
  age_groups = c(
    "Death in the first 24 hours",
    "Early Neonate (24-72 hours)",
    "Early Neonate (72+hrs to 6 days)",
    "Late Neonate (7 to 27 days)")
)
}
