% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{calc_top_tac_pathogens_cc}
\alias{calc_top_tac_pathogens_cc}
\title{Tabulate TAC top pathogens detected and causal chain}
\usage{
calc_top_tac_pathogens_cc(
  d,
  condition,
  specimen_types = NULL,
  specimen_abbrv = NULL,
  age_groups = NULL,
  sites = NULL
)
}
\arguments{
\item{d}{A data object obtained from \code{\link{load_data}}.}

\item{condition}{A string specifying the condition to tabulate causal chain
presence for. To list all possibilities, see
\code{\link{valid_conditions}}.}

\item{specimen_types}{An optional vector of specimen types to include as columns in the table.}

\item{specimen_abbrv}{An optional vector of abbreviations to use for the specified \code{specimen_types}.
calculation. If not provided, all specimen types will be used. See
\code{\link{valid_specimen_types}}.}

\item{age_groups}{Optional vector of age groups.
See \code{\link{valid_age_groups}}.}

\item{sites}{An optional vector of site names to include in the tabulation.
If not provided, all sites will be used. See \code{\link{valid_sites}}.}
}
\description{
Tabulate TAC top pathogens detected and causal chain
}
\examples{
calc_top_tac_pathogens_cc(mock,
  condition = "Lower respiratory infections",
  age_groups = c(
    "Infant (28 days to less than 12 months)",
    "Child (12 months to less than 60 Months)"),
  specimen_types = c(
    "Nasopharyngeal and Oropharyngeal swab",
    "Tissue specimen from lung"),
  specimen_abbrv = c("# NP+", "# Lung+")
)
}
