% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{heatmap_site_age}
\alias{heatmap_site_age}
\title{Plot the percents of a computed CHAMPS table}
\usage{
heatmap_site_age(
  x,
  plot_title = "",
  plot_subtitle = "",
  include_text = "percent"
)
}
\arguments{
\item{x}{A computed CHAMPS object that includes a numerator and
denominator of counts.}

\item{plot_title}{is the title for the plot.}

\item{plot_subtitle}{is the subtitle for the plot.}

\item{include_text}{Whether to include the percentages "percent", or
fraction "fraction". Defaults to NULL which is none.}
}
\description{
Plot the percents of a computed CHAMPS table
}
\examples{
mock_calc <- calc_cc_allcases_by_site_age(mock,
  condition = 'Streptococcus agalactiae')
heatmap_site_age(mock_calc, include_text = "percent",
  plot_title = "TAC Results", plot_subtitle = "Streptococcus pneumoniae")
heatmap_site_age(mock_calc, include_text = "fraction",
  plot_title = "TAC Results", plot_subtitle = "Streptococcus pneumoniae")
}
