% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{html_table_site_age}
\alias{html_table_site_age}
\title{Create an html table of a computed CHAMPS object}
\usage{
html_table_site_age(
  x,
  margin_pct = TRUE,
  margin_denom = FALSE,
  inside_pct = FALSE,
  inside_denom = FALSE
)
}
\arguments{
\item{x}{A computed CHAMPS object that contains a numerator and
denominator of counts.}

\item{margin_pct}{Should percentages be printed in the margins?}

\item{margin_denom}{Should denominators be printed in the margins?}

\item{inside_pct}{Should percentages be printed inside the table?}

\item{inside_denom}{Should denominators be printed inside the table?}
}
\description{
Create an html table of a computed CHAMPS object
}
\examples{
mock_calc <- calc_cc_allcases_by_site_age(mock,
  condition = 'Streptococcus agalactiae')
html_table_site_age(mock_calc)
}
