% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{infectious_causes}
\alias{infectious_causes}
\title{A vector of named infectious causes.}
\format{
An object of class \code{character} of length 13.
}
\usage{
infectious_causes
}
\description{
A vector of named infectious causes.
}
\keyword{datasets}
