% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_load.R
\name{load_data}
\alias{load_data}
\title{Load and pre-process data from relevant CHAMPS database tables}
\usage{
load_data(data_dir)
}
\arguments{
\item{data_dir}{Path to directory containing the de-identified L2 data as
directly downloaded from CHAMPS (see \code{\link{get_l2_data}} for
information on how to access this data). Data files must not be renamed
or modified after downloading.}
}
\description{
Load and pre-process data from relevant CHAMPS database tables
}
\note{
The variables in the data are well-documented in the dataset
  definition PDF file that comes with the L2 data download.
}
