% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\name{\%>\%}
\alias{\%>\%}
\title{Pipe}
\arguments{
\item{lhs, rhs}{an object and a function to apply to it}
}
\description{
Use the pipe function, \code{\%>\%} to turn function composition
into a series of imperative statements.
}
