% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_top_etiol_by_age}
\alias{plot_top_etiol_by_age}
\title{Plot the etiologies of a computed CHAMPS table}
\usage{
plot_top_etiol_by_age(x, etiologies = NULL, top = 6, bar_width = 0.5)
}
\arguments{
\item{x}{A data object obtained from \code{\link{calc_top_etiol_by_age}}.}

\item{etiologies}{The etiologies to include in the graphic with their
shortened names as a named vector. Defaults to NULL. If NULL than the top
variable is used and the names are shortened automatically.}

\item{top}{The number of etiologies to include in the graphic}

\item{bar_width}{defines the size of the bar the chart. Defaults to 0.5.}
}
\description{
Plot the etiologies of a computed CHAMPS table
}
\examples{
mock_top_etiol_by_age <- calc_top_etiol_by_age(mock,
  age_groups = c(
    "Death in the first 24 hours",
    "Early Neonate (24-72 hours)",
    "Early Neonate (72+hrs to 6 days)",
    "Late Neonate (7 to 27 days)"))

plot_top_etiol_by_age(mock_top_etiol_by_age)
}
