% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tac_table}
\alias{tac_table}
\title{A table of TAC name information}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 128 rows and 3 columns.
}
\usage{
tac_table
}
\description{
A table built from the dataset description PDF that comes with the data download.
}
\keyword{datasets}
