% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{valid_conditions}
\alias{valid_conditions}
\title{Get a list of valid conditions}
\usage{
valid_conditions(d)
}
\arguments{
\item{d}{A data object obtained from \code{\link{load_data}}.}
}
\description{
Get a list of valid conditions
}
\examples{
\dontrun{
d <- load_data(...)
valid_conditions(d)
}
}
