% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{valid_icds}
\alias{valid_icds}
\title{Get a list of DeCoDe "Other Significant Cause" ICD 10 codes that exist in the
  CHAMPS data}
\usage{
valid_icds(d)
}
\arguments{
\item{d}{A data object obtained from \code{\link{load_data}}.}
}
\description{
Get a list of DeCoDe "Other Significant Cause" ICD 10 codes that exist in the
  CHAMPS data
}
\examples{
\dontrun{
d <- load_data(...)
valid_icds(d)
}
}
