% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{valid_pathogens}
\alias{valid_pathogens}
\title{Get a list of valid TAC results pathogens based on the provided CHAMPS data}
\usage{
valid_pathogens()
}
\description{
Get a list of valid TAC results pathogens based on the provided CHAMPS data
}
\examples{
\dontrun{
d <- load_data(...)
valid_pathogens()
}
}
