% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{valid_sites}
\alias{valid_sites}
\title{Get a list of valid site names based on the provided CHAMPS data}
\usage{
valid_sites(d)
}
\arguments{
\item{d}{A data object obtained from \code{\link{load_data}}.}
}
\description{
Get a list of valid site names based on the provided CHAMPS data
}
\examples{
\dontrun{
d <- load_data(...)
valid_sites(d)
}
}
