% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{valid_specimen_types}
\alias{valid_specimen_types}
\title{Get a list of valid TAC results specimen types based on the provided
  CHAMPS data}
\usage{
valid_specimen_types(d)
}
\arguments{
\item{d}{A data object obtained from \code{\link{load_data}}.}
}
\description{
Get a list of valid TAC results specimen types based on the provided
  CHAMPS data
}
\examples{
\dontrun{
d <- load_data(...)
valid_specimen_types(d)
}
}
